Features in top js file

+++ app/Http/Controllers/FriendController.php
@@ -7,6 +7,7 @@ use App\Models\Notification;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class FriendController extends Controller
 {
@@ -16,38 +17,49 @@ class FriendController extends Controller
      * @return \Illuminate\Http\Response
      */
     
-    public function index()
+    public function index($search=false)
     {
-        // $friend_users=User::with('friend')->find(Auth::id());
-        $friend_users=Auth::user()->friend()->with('user2')->where('acceptable',1)->get() ;
-        $friend_users = $friend_users->map(function ($query) {
-           return $query->user2;
-         });
-        return response()->json($friend_users);
+      
+        $friend_users= DB::table('friends')->where([['user1_id',Auth::id()],['acceptable',1]])->orWhere([['user2_id',Auth::id()],['acceptable',1]])->get(['id','user1_id','user2_id']);
+
+        $id=[];
+            foreach($friend_users as $u )
+            { 
+                if($u->user1_id != Auth::id())
+                array_push($id,$u->user1_id );
+                else
+                 array_push($id,$u->user2_id );
+            }
+            if($search)
+            return $id; 
+
+            $users=DB::table('users')->whereIn('id',$id)->get(['id','name','img']);
+            return response()->json($users);
+    }
+    public function search_friends(Request $request){
+       $users_id= $this->index(true);
+       $users=DB::table('users')->whereIn('id',$users_id)->where('name', 'like', "%" . $request->name . "%")->get(['id','name','img']);
+
+       return $users;
     }
 
-    // public function getfriends(){
-        
-    //     $friends=Friend::where('user1_id',Auth::id())->where('acceptable',1)->get();
-    //     return $friends;
+  
 
-    // }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
-        $friend=Friend::where('user1_id',Auth::id())->where('user2_id',$request->user_id)->first() ;
-    //  return $friend ; 
+        if($request->user_id==Auth::id())
+        {
+            return response()->json(['message '=>'send add to your self ','status'=>0],200);
+        }
+        $friend=Friend::where([['user1_id',$request->user_id],['user2_id',Auth::id()]])->
+                        orWhere([['user1_id',Auth::id()],['user2_id',$request->user_id]])->
+                        first();
+
         if($friend ==null )
         {
-            $user=User::find($request->user_id);
+            
             $f= Friend::create(['user1_id'=>Auth::id(),'user2_id'=>$request->user_id]);
-            $f= Friend::create(['user1_id'=>$request->user_id,'user2_id'=>Auth::id()]);
+         
             Notification::create(['refernce'=>$f->id,'owner_id'=>$request->user_id , 'user_id'=>Auth::id(),'title'=>Auth::user()->name ,'body'=>'Send you a friend request.' ,'type'=>'request']);
             return response()->json([
                 'id'=>$f->id,   
@@ -55,67 +67,51 @@ class FriendController extends Controller
                 'message'=>'send add friend  successfuly'
             ],200);
         }
-        return response()->json([
-            // 'id'=>$f->id,   
+        return response()->json([  
             'status'=>0,
             'message'=>'allready send add friend successfuly'
         ],200);
         
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
+  
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request,$frindship_id)
+    public function update($frindship_id)
     {
-        // return'  '.$frindship_id;
 
         $friendship=Friend::find($frindship_id);
-        // return $friendship;
-        $fr= Friend::where('user1_id',$friendship->user2_id)->where('user2_id',$friendship->user1_id)->first();
-        $notification= Notification::find($request->notification_id);
+        if($friendship==null)
+        return response()->json(['message'=>'not found','status'=>0],200);
+        
+        if($friendship->acceptable==1)
+        return response()->json(['message'=>'alleady accepted','status'=>0],200);
 
-        $notification->delete();
+
         $friendship->update(['acceptable'=>1]);
-        $fr->update(['acceptable'=>1]);
+
+        $notification= Notification::where('refernce',$frindship_id)->first();
+        $notification->delete();
+
+      
         return response()->json([
             'status'=>1,
             'message'=>'added successfuly'
         ],200);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($frindship_id , Request $request)
+
+    public function destroy($frindship_id )
     {
         $friendship=Friend::find($frindship_id);
-
-        // return($request->notification_id).'1';
-        $notification= Notification::find($request->notification_id);
-        $fr= Friend::where('user1_id',$friendship->user2_id)->where('user2_id',$friendship->user1_id)->first();
-
-        $notification->delete();
+        if($friendship==null)
+        return response()->json(['message'=>'not found','status'=>0],200);
+        
         $friendship->delete();
-        $fr->delete();
+
+        $notification= Notification::where('refernce',$frindship_id)->first();
+        if($notification)
+        $notification->delete();
+        
 
         return response()->json([
             'status'=>1,
